Regressed from uncompilable code. Simplified design.

+++ include/FalconEngine/Graphics/Scenes/Node.inl
@@ -1,4 +1,4 @@
-inline int Node::ChildrenNum() const
+inline size_t Node::ChildrenNum() const
 {
     return m_children.size();
 }
